fixed coupon issue fixed price issue
