+++ resources/views/pages/shop.blade.php
@@ -72,10 +72,10 @@
                                         <h2><a href="{{ url('/product/details/'.$row->id) }}">{{$row->product_name}}</a></h2>
                                         <ul class="product__price">
                                              @if($row->discount_price == NULL)
-                                            <li >GH₵ {{$row->selling_price }}</li>
+                                            <li >GH₵ {{$row->selling_price / 100}}</li>
                                               @else
-                                              <li class="new__price">GH₵ {{$row->discount_price }}</li>
-                                            <li class="old__price">GH₵ {{$row->selling_price }}</li>
+                                              <li class="new__price">GH₵ {{$row->discount_price / 100}}</li>
+                                            <li class="old__price">GH₵ {{$row->selling_price / 100}}</li>
                                               @endif
                                         </ul>
                                     </div>
